# LG-3769: Optimize image assets