--- a/Makefile
+++ b/Makefile
@@ -56,7 +56,9 @@ normalize_yaml:
 	find ./config/locales -type f | xargs ./scripts/normalize-yaml config/country_dialing_codes.yml
 
 optimize_svg:
-	find app/assets/images public -name '*.svg' | xargs ./node_modules/.bin/svgo --multipass --config '{"plugins":[{"removeAttrs":{"attrs":"data-name"}}]}'
+	# Without disabling minifyStyles, keyframes are removed (e.g. `app/assets/images/id-card.svg`).
+	# See: https://github.com/svg/svgo/issues/888
+	find app/assets/images public -name '*.svg' | xargs ./node_modules/.bin/svgo --multipass --disable minifyStyles --config '{"plugins":[{"removeAttrs":{"attrs":"data-name"}}]}'
 
 optimize_assets: optimize_svg