--- a/app/assets/images/alert/error.svg
+++ b/app/assets/images/alert/error.svg
@@ -1,1 +1,1 @@
-<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16"><title>error</title><g id="Layer_2" data-name="Layer 2"><g id="Layer_1-2" data-name="Layer 1"><g id="Layer_2-2" data-name="Layer 2"><g id="Layer_1-2-2" data-name="Layer 1-2"><circle cx="8" cy="8" r="8" fill="#eb4d67"/><path d="M11.9,6.6H4.1a1.6,1.6,0,1,0,0,3.2h7.8a1.6,1.6,0,1,0,0-3.2Z" fill="#fff"/></g></g></g></g></svg>
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16"><circle cx="8" cy="8" r="8" fill="#eb4d67"/><path d="M11.9 6.6H4.1a1.6 1.6 0 100 3.2h7.8a1.6 1.6 0 100-3.2z" fill="#fff"/></svg>