--- a/app/assets/images/close-white-alt.svg
+++ b/app/assets/images/close-white-alt.svg
@@ -1,3 +1,1 @@
-<svg xmlns="http://www.w3.org/2000/svg" width="15" height="15" viewBox="0 0 15 15">
-  <path fill="none" stroke="#FFF" stroke-linecap="round" stroke-width="2" d="M0 13.0332964L13.0332964 0M13.0332964 13.0332964L0 0" transform="translate(1 1)"/>
-</svg>
+<svg xmlns="http://www.w3.org/2000/svg" width="15" height="15"><path fill="none" stroke="#FFF" stroke-linecap="round" stroke-width="2" d="M1 14.033L14.033 1m0 13.033L1 1"/></svg>