--- a/app/assets/images/minus.svg
+++ b/app/assets/images/minus.svg
@@ -1,8 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- Generator: Adobe Illustrator 21.0.2, SVG Export Plug-In . SVG Version: 6.00 Build 0)  -->
-<svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-	 viewBox="0 0 402 402" enable-background="new 0 0 402 402" xml:space="preserve">
-<path fill="#0071BB" d="M394,154.2c-5.3-5.3-11.8-8-19.4-8H27.4c-7.6,0-14.1,2.7-19.4,8c-5.3,5.3-8,11.8-8,19.4v54.8
-	c0,7.6,2.7,14.1,8,19.4c5.3,5.3,11.8,8,19.4,8h347.2c7.6,0,14.1-2.7,19.4-8c5.3-5.3,8-11.8,8-19.4v-54.8
-	C402,166,399.3,159.5,394,154.2L394,154.2z"/>
-</svg>
+<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 402 402"><path fill="#0071BB" d="M394 154.2c-5.3-5.3-11.8-8-19.4-8H27.4c-7.6 0-14.1 2.7-19.4 8-5.3 5.3-8 11.8-8 19.4v54.8c0 7.6 2.7 14.1 8 19.4 5.3 5.3 11.8 8 19.4 8h347.2c7.6 0 14.1-2.7 19.4-8 5.3-5.3 8-11.8 8-19.4v-54.8c0-7.6-2.7-14.1-8-19.4z"/></svg>