--- a/package.json
+++ b/package.json
@@ -62,6 +62,7 @@
     "prettier": "^2.2.1",
     "react-test-renderer": "^17.0.1",
     "sinon": "^9.2.2",
+    "svgo": "^1.3.2",
     "typescript": "^4.1.3",
     "webpack-dev-server": "^3.11.1"
   }