… app/assets/images/2FA-voice.svg
… app/assets/images/alert/fail-lg.svg
… app/assets/images/alert/fail-x.svg
… app/assets/images/alert/forgot.svg
… app/assets/images/alert/icon-lock-alert-important.svg
… app/assets/images/alert/success.svg
… app/assets/images/alert/temp-lock.svg
… app/assets/images/alert/unphishable.svg
… app/assets/images/alert/warning-lg.svg
… app/assets/images/alert/warning.svg
… app/assets/images/carat-right.svg
… app/assets/images/check-email.svg
… app/assets/images/clock.svg
… app/assets/images/come-back.svg
… app/assets/images/get-started/2FA.svg
… app/assets/images/get-started/ID.svg
… app/assets/images/get-started/email-password.svg
… app/assets/images/get-started/financial.svg
… app/assets/images/get-started/personal-details.svg
… app/assets/images/globe-blue.svg
… app/assets/images/globe-white.svg
… app/assets/images/ico-copy.svg
… app/assets/images/ico-download.svg
… app/assets/images/ico-print.svg
… app/assets/images/ico-refresh.svg
… app/assets/images/icon-dot-gov.svg
… app/assets/images/icon-https.svg
… app/assets/images/icon-lock-alert-important.svg
… app/assets/images/id-card.svg
… app/assets/images/idv/bar_code.png