**Why**: As a user, I want images to be as small file size as possible, so that pages load fast.

**Impact:**

<details><summary>Click to expand SVG results</summary>
webauthn-verified.svg:
Done in 23 ms!
0.813 KiB - 45.4% = 0.444 KiB

state-id-round.svg:
Done in 57 ms!
6.769 KiB - 0.5% = 6.735 KiB

scissors.svg:
Done in 13 ms!
1.394 KiB - 12.7% = 1.217 KiB

globe-white.svg:
Done in 9 ms!
1.124 KiB - 13.6% = 0.972 KiB

ico-copy.svg:
Done in 4 ms!
0.53 KiB - 21.5% = 0.416 KiB

2FA-voice.svg:
Done in 25 ms!
4.48 KiB - 4.7% = 4.269 KiB

pkey-block.svg:
Done in 6 ms!
0.825 KiB - 15.5% = 0.697 KiB

warning.svg:
Done in 4 ms!
0.589 KiB - 22.4% = 0.457 KiB

shield.svg:
Done in 2 ms!
0.42 KiB - 25.8% = 0.312 KiB

user.svg:
Done in 2 ms!
0.449 KiB - 43% = 0.256 KiB

email.svg:
Done in 9 ms!
1.153 KiB - 17.1% = 0.956 KiB

close-white-alt.svg:
Done in 1 ms!
0.245 KiB - 29.1% = 0.174 KiB

globe-blue.svg:
Done in 5 ms!
1.126 KiB - 13.4% = 0.975 KiB

ico-refresh.svg:
Done in 7 ms!
0.752 KiB - 19.6% = 0.604 KiB

verified.svg:
Done in 4 ms!
0.795 KiB - 39.3% = 0.482 KiB

icon-lock-alert-important.svg:
Done in 9 ms!
2.405 KiB - 58.7% = 0.994 KiB

check-email.svg:
Done in 17 ms!
2.083 KiB - 15.8% = 1.753 KiB

lock.svg:
Done in 5 ms!
0.444 KiB - 43.3% = 0.252 KiB

warning.svg:
Done in 2 ms!
0.513 KiB - 38.7% = 0.314 KiB

icon-lock-alert-important.svg:
Done in 4 ms!
2.405 KiB - 58.7% = 0.994 KiB

forgot.svg:
Done in 36 ms!
2.093 KiB - 15.3% = 1.772 KiB

fail-lg.svg:
Done in 60 ms!
1.548 KiB - 18.4% = 1.263 KiB

temp-lock.svg:
Done in 6 ms!
1.828 KiB - 24.6% = 1.378 KiB

fail-x.svg:
Done in 8 ms!
1.567 KiB - 16.9% = 1.302 KiB

success.svg:
Done in 2 ms!
0.462 KiB - 42.5% = 0.266 KiB

warning-lg.svg:
Done in 9 ms!
1.67 KiB - 17.1% = 1.385 KiB

error.svg:
Done in 2 ms!
0.373 KiB - 50.8% = 0.184 KiB

unphishable.svg:
Done in 6 ms!
0.943 KiB - 46.3% = 0.507 KiB

ico-download.svg:
Done in 37 ms!
2.89 KiB - 66.5% = 0.967 KiB

plus.svg:
Done in 10 ms!
0.755 KiB - 4% = 0.725 KiB

icon-dot-gov.svg:
Done in 5 ms!
0.781 KiB - 11.9% = 0.688 KiB

piv-cac-round.svg:
Done in 52 ms!
6.773 KiB - 0.4% = 6.744 KiB

security-key.svg:
Done in 56 ms!
7.033 KiB - 13.8% = 6.06 KiB

come-back.svg:
Done in 63 ms!
6.463 KiB - 24% = 4.909 KiB

square-gsa-dark.svg:
Done in 7 ms!
3.102 KiB - 67% = 1.024 KiB

square-gsa.svg:
Done in 10 ms!
2.191 KiB - 55.8% = 0.969 KiB

18f.svg:
Done in 10 ms!
1.341 KiB - 32.7% = 0.902 KiB

generic.svg:
Done in 3 ms!
1.167 KiB - 51.4% = 0.567 KiB

user-signup-ial1.svg:
Done in 49 ms!
20.646 KiB - 62.4% = 7.768 KiB

clock.svg:
Done in 17 ms!
2.327 KiB - 24.4% = 1.759 KiB

ico-print.svg:
Done in 3 ms!
0.591 KiB - 20.7% = 0.469 KiB

icon-https.svg:
Done in 3 ms!
0.396 KiB - 4.9% = 0.377 KiB

user-signup-ial2.svg:
Done in 73 ms!
22.466 KiB - 65.9% = 7.665 KiB

id-card.svg:
Done in 21 ms!
3.692 KiB - 64.6% = 1.307 KiB

user-access.svg:
Done in 80 ms!
23.18 KiB - 30% = 16.217 KiB

sign-in.svg:
Done in 2 ms!
0.625 KiB - 32.5% = 0.422 KiB

p-key.svg:
Done in 4 ms!
1.289 KiB - 18% = 1.058 KiB

logo.svg:
Done in 10 ms!
2.189 KiB - 24.2% = 1.659 KiB

carat-right.svg:
Done in 2 ms!
0.446 KiB - 27.8% = 0.322 KiB

financial.svg:
Done in 15 ms!
2.205 KiB - 19.1% = 1.783 KiB

email-password.svg:
Done in 8 ms!
3.713 KiB - 47% = 1.968 KiB

2FA.svg:
Done in 18 ms!
2.776 KiB - 19.5% = 2.236 KiB

personal-details.svg:
Done in 8 ms!
1.995 KiB - 13.8% = 1.721 KiB

ID.svg:
Done in 8 ms!
5.064 KiB - 54.5% = 2.305 KiB

minus.svg:
Done in 2 ms!
0.597 KiB - 49.9% = 0.299 KiB

logo-white.svg:
Done in 9 ms!
2.169 KiB - 24.7% = 1.633 KiB

safari-pinned-tab.svg:
Done in 8 ms!
1.111 KiB - 33.9% = 0.734 KiB

logo.svg:
Done in 20 ms!
5.155 KiB - 68.1% = 1.644 KiB

logo-white.svg:
Done in 22 ms!
5.155 KiB - 68.6% = 1.617 KiB
</details>

<details><summary>Click to expand PNG results</summary>
<pre>app/assets/images/idv/bar_code.png 2.74kb 2.41kb -12%
app/assets/images/idv/phone.png 5.89kb 1.91kb -68%
app/assets/images/idv/switch.png 5.21kb 1.0kb -81%
app/assets/images/logo.png 6.34kb 3.61kb -43%
app/assets/images/state-id-confirm@3x.png 17.86kb 14.55kb -19%
app/assets/images/us-flag.png 0.28kb 0.13kb -54%
public/apple-touch-icon.png 3.81kb 3.02kb -21%
public/favicon-16x16.png 0.91kb 0.56kb -38%
public/favicon-32x32.png 1.33kb 0.93kb -32%
public/mstile-150x150.png 3.66kb 3.11kb -15%</pre>
</details>